docker/[local/remote] support

+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"context"
 	"errors"
 	"flag"
 	"fmt"
@@ -11,6 +12,10 @@ import (
 	"strings"
 
 	"github.com/abiosoft/ishell"
+	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/container"
+	"github.com/docker/docker/client"
+	"github.com/docker/go-connections/nat"
 	"github.com/hashicorp/hcl"
 	"github.com/hpcloud/tail"
 	"github.com/tevino/abool"
@@ -31,6 +36,7 @@ func main() {
 	}
 
 	shell := ishell.New()
+	shell.SetHomeHistoryPath(".ishell_history/glorious")
 
 	// display welcome info.
 	shell.Println(banner)
@@ -209,6 +215,7 @@ type Status struct {
 	CurrentStatus UnitStatus
 	Cmd           *exec.Cmd
 	OutFile       *os.File
+	CurrentSlot   *Slot
 
 	shutdownRequested *abool.AtomicBool
 }
@@ -258,12 +265,181 @@ func (u *Unit) OutputFile() (*os.File, error) {
 	return f, err
 }
 
+func (u *Unit) identifySlot() (*Slot, error) {
+	// Short circuit if only a single slot exists.
+	if len(u.Slots) == 0 {
+		return nil, errors.New("unit has no slots")
+	} else if len(u.Slots) == 1 {
+		return &(u.Slots[0]), nil
+	}
+
+	// Always first identify the default slot
+	var defaultSlot *Slot
+	var resolverResults = make([]bool, len(u.Slots))
+	for i, slot := range u.Slots {
+		if slot.IsDefault() {
+			if defaultSlot != nil {
+				// Two slots are defined as default, result: barf.
+				return nil, errors.New("there can only be one default slot")
+			}
+			defaultSlot = &slot
+		}
+
+		val, err := slot.Resolve(u)
+		if err != nil {
+			return nil, err
+		}
+		resolverResults[i] = val
+
+	}
+
+	// If no slot is defined as the default, grab the first one.
+	defaultSlot = &(u.Slots[0])
+
+	// Run through resolvers. Take the last resolved value if any.
+	var resolvedSlot *Slot
+	for i := len(resolverResults) - 1; i >= 0; i-- {
+		if resolverResults[i] {
+			resolvedSlot = &(u.Slots[i])
+			break
+		}
+	}
+
+	if resolvedSlot != nil {
+		return resolvedSlot, nil
+	}
+	return defaultSlot, nil
+}
+
 func (u *Unit) Start(ctxt *ishell.Context) error {
 	// now for some tomfoolery
-	slot := u.Slots[0]
-	pieces := strings.Split(slot.Cmd, " ")
+	slot, err := u.identifySlot()
+	if err != nil {
+		return err
+	}
+
+	return slot.Start(u, ctxt)
+}
+
+func (s *Slot) Start(u *Unit, ctxt *ishell.Context) error {
+	providerType := s.Provider.Type
+	if len(providerType) == 0 {
+		return errors.New("no provider given")
+	}
+
+	switch providerType {
+	case "bash/local":
+		return s.startBashLocal(u, ctxt)
+	case "docker/local":
+		return s.startDockerLocal(u, ctxt)
+	case "docker/remote":
+		return s.startDockerRemote(u, ctxt)
+	default:
+		return errors.New("unknown provider")
+	}
+}
+
+func (s *Slot) startDockerLocal(u *Unit, ctxt *ishell.Context) error {
+	return s.startDockerInternal(u, ctxt, false)
+}
+
+func (s *Slot) startDockerRemote(u *Unit, ctxt *ishell.Context) error {
+	return s.startDockerInternal(u, ctxt, true)
+}
+
+func (s *Slot) startDockerInternal(u *Unit, ctxt *ishell.Context, remote bool) error {
+	image := s.Provider.Image
+	if len(image) == 0 {
+		return errors.New("no image provided")
+	}
+
+	options := []client.Opt{
+		client.FromEnv,
+		client.WithAPIVersionNegotiation(),
+	}
+	if remote {
+		options = append(options, client.WithHost(s.Provider.Remote.Host))
+	}
+
+	ctx := context.Background()
+	cli, err := client.NewClientWithOpts(options...)
+	if err != nil {
+		return err
+	}
+
+	// first see if the image exists
+	_, _, err = cli.ImageInspectWithRaw(ctx, image)
+	if err != nil {
+		if client.IsErrNotFound(err) {
+			ctxt.Println("\nimage not found locally, trying to pull...")
+			_, pullErr := cli.ImagePull(
+				ctx,
+				image,
+				types.ImagePullOptions{},
+			)
+			if pullErr != nil {
+				return pullErr
+			}
+		} else {
+			return err
+		}
+	}
+
+	hostConfig := &container.HostConfig{}
+	if len(s.Provider.Ports) > 0 {
+		bindings := nat.PortMap{}
+		for _, port := range s.Provider.Ports {
+			vals, err := nat.ParsePortSpec(port)
+			if err != nil {
+				return err
+			}
+			for _, val := range vals {
+				bindings[val.Port] = []nat.PortBinding{val.Binding}
+			}
+		}
+		hostConfig.PortBindings = bindings
+	}
+
+	resp, err := cli.ContainerCreate(ctx, &container.Config{
+		Image: image,
+	}, hostConfig, nil, u.Name)
+	if err != nil {
+		return err
+	}
+
+	if err := cli.ContainerStart(
+		ctx,
+		resp.ID,
+		types.ContainerStartOptions{},
+	); err != nil {
+		return err
+	}
+
+	u.Status = &Status{
+		CurrentStatus: Running,
+		CurrentSlot:   s,
+
+		shutdownRequested: abool.New(),
+	}
+	u.Status.shutdownRequested.UnSet()
+
+	ctxt.Println("begun as container ", resp.ID)
+
+	return nil
+}
+
+func (s *Slot) startBashLocal(u *Unit, ctxt *ishell.Context) error {
+	cmd := s.Provider.Cmd
+	if len(cmd) == 0 {
+		return errors.New("no `cmd` provided")
+	}
+
+	// we don't care if it's set or not
+	workingDir := s.Provider.WorkingDir
+
+	pieces := strings.Split(cmd, " ")
 	c := exec.Cmd{}
-	c.Dir = slot.WorkingDir
+	c.Dir = workingDir
 	c.Path = pieces[0]
 	if len(pieces) > 1 {
 		c.Args = pieces[1:]
@@ -285,6 +461,7 @@ func (u *Unit) Start(ctxt *ishell.Context) error {
 		CurrentStatus: Running,
 		Cmd:           &c,
 		OutFile:       outputFile,
+		CurrentSlot:   s,
 
 		shutdownRequested: abool.New(),
 	}
@@ -312,17 +489,57 @@ func (u *Unit) Stop(ctxt *ishell.Context) error {
 
 	u.Status.shutdownRequested.Set()
 
-	if err := u.Status.Cmd.Process.Kill(); err != nil {
+	// TODO(ttacon): move to refactored function
+	if u.Status.CurrentSlot.Provider.Type == "bash/local" {
+
+		if err := u.Status.Cmd.Process.Kill(); err != nil {
+			return err
+		}
+
+		if err := u.Status.OutFile.Close(); err != nil {
+			return err
+		}
+
+		u.Status.Cmd.Stdout = nil
+		u.Status.Cmd.Stderr = nil
+	} else if u.Status.CurrentSlot.Provider.Type == "docker/local" {
+		return u.Status.CurrentSlot.stopDocker(u, ctxt, false)
+	} else if u.Status.CurrentSlot.Provider.Type == "docker/remote" {
+		return u.Status.CurrentSlot.stopDocker(u, ctxt, true)
+	} else {
+		return errors.New("unknown provider for unit, cannot stop, also probably wasn't started")
+	}
+
+	return nil
+}
+
+func (s *Slot) stopDocker(u *Unit, ctxt *ishell.Context, remote bool) error {
+	options := []client.Opt{
+		client.FromEnv,
+		client.WithAPIVersionNegotiation(),
+	}
+
+	if remote {
+		options = append(options, client.WithHost(s.Provider.Remote.Host))
+	}
+
+	ctx := context.Background()
+	cli, err := client.NewClientWithOpts(options...)
+	if err != nil {
 		return err
 	}
 
-	if err := u.Status.OutFile.Close(); err != nil {
+	if err := cli.ContainerStop(ctx, u.Name, nil); err != nil {
 		return err
 	}
 
-	u.Status.Cmd.Stdout = nil
-	u.Status.Cmd.Stderr = nil
-
+	if err := cli.ContainerRemove(
+		ctx,
+		u.Name,
+		types.ContainerRemoveOptions{},
+	); err != nil {
+		return err
+	}
 	return nil
 }
 
@@ -337,19 +554,38 @@ func (u *Unit) Tail(ctxt *ishell.Context) error {
 	}
 
 	return nil
-
-	/*
-		// let's be cheeky for now and smarter later (implementing `tail`)
-		data, err := exec.Command("tail", "-100", u.Status.OutFile.Name()).Output()
-		if err != nil {
-			return err
-		}
-		ctxt.Println(string(data))
-		return nil
-	*/
 }
 
 type Slot struct {
-	Cmd        string `hcl:"cmd"`
+	Provider *Provider         `hcl:"provider"`
+	Resolver map[string]string `hcl:"resolver"`
+}
+
+type Provider struct {
+	Type string `hcl:"type"`
+
 	WorkingDir string `hcl:"workingDir"`
+	Cmd        string `hcl:"dir"`
+
+	Image string   `hcl:"image"`
+	Ports []string `hcl:"ports"`
+
+	Remote RemoteInfo `hcl:"remote"`
+}
+
+type RemoteInfo struct {
+	Host         string `hcl:"host"`
+	User         string `hcl:"user"`
+	IdentityFile string `hcl:"identityFile"`
+}
+
+func (s Slot) IsDefault() bool {
+	typ, ok := s.Resolver["type"]
+
+	return ok && typ == "default"
+}
+
+func (s Slot) Resolve(u *Unit) (bool, error) {
+
+	return false, nil
 }
